Update of the structure of the project
