+++ src/agenda/controllers/agenda.controller.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Get, Post } from '@nestjs/common';
-import { AgendaService } from '../../../services/agenda.service';
-import { Agenda } from '../../../entities/agenda.entity';
+import { AgendaService } from '../services/agenda.service';
+import { Agenda } from '../entities/agenda.entity';
 
 @Controller('agenda')
 export class AgendaController {
