+++ src/auth/auth.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+
+import { AuthController } from './controllers/auth.controller';
+import { AuthService } from './services/auth.service';
+
+@Module({
+  controllers: [AuthController],
+  providers: [AuthService],
+})
+export class AuthModule {}
