+++ src/auth/controllers/auth.controller.ts
@@ -1,14 +1,10 @@
-import { Body, Controller, Delete, Get, HttpCode, HttpStatus, Param, ParseBoolPipe, ParseIntPipe, Patch, Post, Put, Query, Req, Res } from '@nestjs/common';
+import { Body, Controller, HttpStatus, Post, Res } from '@nestjs/common';
 import { Response } from 'express';
-import { AuthService } from '../../services/auth.service'
-import { AuthPipe } from '../../common/pipes/auth/auth.pipe';
-import { AuthDto } from '../../dtos/auth.dto';
-
+import { AuthService } from '../services/auth.service';
 
 @Controller('autenticacion')
 export class AuthController {
-
-    constructor(private readonly AuthService: AuthService) { }
+  constructor(private readonly AuthService: AuthService) {}
 
   @Post('/login')
   async funcionAutenticar(@Body() data: any, @Res() res: Response) {
@@ -17,7 +13,7 @@ export class AuthController {
     //const result = await this.appAuthService.login(data);
 
     // Verificar si la autenticación fue exitosa
-    const result = await this.AuthService.login(data);
+    const result = await this.AuthService.login();
 
     // Verificar si la autenticación fue exitosa
     if (result) {
@@ -25,7 +21,7 @@ export class AuthController {
         message: 'Login successful',
         data: result, // Puedes enviar tokens o datos de sesión aquí
         token: '123456789',
-        status: 200
+        status: 200,
       });
     } else {
       return res.status(HttpStatus.UNAUTHORIZED).json({
@@ -33,5 +29,4 @@ export class AuthController {
       });
     }
   }
-  
 }
