+++ src/auth/services/auth.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AppAuthService } from './auth.service';
+import { AuthService } from './auth.service';
 
 describe('ServicesService', () => {
-  let service: AppAuthService;
+  let service: AuthService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [AppAuthService],
+      providers: [AuthService],
     }).compile();
 
-    service = module.get<AppAuthService>(AppAuthService);
+    service = module.get<AuthService>(AuthService);
   });
 
   it('should be defined', () => {
