+++ src/common/pipes/auth/auth.pipe.ts
@@ -1,14 +1,19 @@
-import { ArgumentMetadata, HttpException, HttpStatus, Injectable, PipeTransform } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Injectable,
+  PipeTransform,
+} from '@nestjs/common';
 
 @Injectable()
 export class AuthPipe implements PipeTransform {
-  transform(value: any, metadata: ArgumentMetadata) {
+  transform(value: any) {
     console.log(value);
-    
+
     const ageNumber = parseInt(value.age.toString(), 10);
     if (isNaN(ageNumber)) {
-      throw new HttpException('Age must be a number',HttpStatus.BAD_REQUEST);
+      throw new HttpException('Age must be a number', HttpStatus.BAD_REQUEST);
     }
-    return { ...value, age:ageNumber};
+    return { ...value, age: ageNumber };
   }
 }
