--- src/controllers/auth/auth.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { AuthController } from './auth.controller';
-import { AuthService } from '../../services/auth.service';
-
-@Module({
-    controllers: [AuthController],
-    providers: [AuthService],
-})
-export class AuthModule {}
