--- src/dtos/auth.dto.ts
@@ -1,15 +1,0 @@
-import { IsNotEmpty, IsString } from "class-validator"
-
-export class AuthDto {
-    @IsString()
-    @IsNotEmpty()
-    usuario:string
-
-    @IsString()
-    @IsNotEmpty()
-    password:string
-
-    @IsString()
-    @IsNotEmpty()
-    sistema:string
-}
